# mnt/private/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "description": "Documentation de l'API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "API de développement"
        }
    ],
    "paths": {
...
        "/api/evaluations/history/{userId}/{clubId}": {
            "get": {
                "tags": [
                    "Evaluations"
                ],
                "summary": "Displays the evaluation history for a given user and club",
                "operationId": "bf73df24555cdfe25eb2caa5fe5ec4cc",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "The ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "clubId",
                        "in": "path",
                        "description": "The ID of the club",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evaluation history retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Evaluation"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User or club not found"
                    }
                }
            }
        },
        "/api/evaluations/search": {
            "get": {
                "tags": [
                    "Evaluations"
                ],
                "summary": "Search for evaluations based on various filters",
                "operationId": "0b014d6a66511bb52ed62a98e67be217",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "Filter by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Evaluation"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/evaluations/update": {
            "post": {
                "tags": [
                    "Evaluations"
                ],
                "summary": "Update or create an evaluation",
                "operationId": "fa30472c1d58c73a5136fba6b36f8ac0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "eval_id",
                                    "statut_id",
                                    "user_id",
                                    "abi_id",
                                    "sess_id"
                                ],
                                "properties": {
                                    "eval_id": {
                                        "type": "integer"
                                    },
                                    "statut_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "abi_id": {
                                        "type": "integer"
                                    },
                                    "sess_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evaluation updated or created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nouvelle évaluation créée avec succès!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Evaluation": {
                "required": [
                    "USER_ID",
                    "STATUSTYPE_ID",
                    "SESS_ID",
                    "ABI_ID"
                ],
                "properties": {
                    "id": {
                        "description": "Evaluation ID",
                        "type": "integer"
                    },
                    "USER_ID": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "STATUSTYPE_ID": {
                        "description": "Status Type ID",
                        "type": "integer"
                    },
                    "SESS_ID": {
                        "description": "Session ID",
                        "type": "integer"
                    },
                    "ABI_ID": {
                        "description": "Ability ID",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}